Change listeners to behave state
Added a state machine to the stage director to handle when the full
screen click listener was active. If the menu is up, I don't want the
user clicking on the screen and not the menu options. It would make the
story progress even if there was no option selected.